* Make the postprocessing level float since different codecs have different   ranges
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@2265 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/include/avdec_private.h
@@ -752,9 +752,9 @@ struct bgav_options_s
   int seamless;
   int seek_subtitles;
 
-  /* Postprocessing level (0..6) */
+  /* Postprocessing level (0.0 .. 1.0) */
   
-  int pp_level;
+  float pp_level;
 
   char * dvb_channels_file;
